v2.33.57 round 11: OceanMap internals split (606→303 LOC, -50%)
backlog #130 — OceanMap.tsx 606 LOC monolith 拆 1 type + 1 helper + 3
hook + 1 compose shell。0 UX change，純內部架構重組。

REFACTOR (new files):
- src/lib/mapTypes.ts (31 LOC) — MapPin / MapPinType / Coord pure type
- src/lib/mapHelpers.ts (239 LOC) — markerStyle / markerContent /
  segmentStyle / buildSegments pure helper
- src/hooks/useMapMarkers.ts (120 LOC) — markersRef + prevFocusRef +
  2 effect (marker lifecycle / focus diff-update)
- src/hooks/useMapViewport.ts (88 LOC) — fitDoneRef + 3 effect
  (fit / resize / pan)
- src/hooks/useMapSegments.ts (30 LOC) — SegmentPair[] useMemo

OceanMap.tsx 606 → 303 (compose shell + Segment subcomponent + JSX).
Hook call 順序對齊原 effect 順序 (markers → viewport → segments).

BACKWARD COMPAT (17+ caller 不動):
- OceanMap.tsx re-export markerStyle/markerContent/buildSegments
- useMapData.ts re-export MapPin/MapPinType
- useRoute.ts re-export Coord
- 拆 mapTypes 同時解掉 lib→hooks reverse import

ARCHITECTURAL GUARD:
- tests/unit/round-11-oceanmap-split.test.ts — 9 個 source-grep
  (lib leaf / 3 hook 單一職責 / compose shell / call order / re-export)
- lib-no-reverse-import.test.ts 仍綠
- 既有 4 個 OceanMap test 全綠 (33 pass)
- 2 個 source-grep test 路徑更新

2403/2403 全綠 (+9). tsc clean.

closes #130 + 完整收尾 #124 (Round 6c).

doc: docs/code-review/round-11-oceanmap-split.md
